+++ CHANGELOG.md
@@ -3,10 +3,16 @@
 ## unreleased
 
 * PayPal
-    * Collect device information to improve PayPal app switch eligibility when it's enabled
     * Add `PayPalCheckoutRequest.campaigns` to associate campaigns with a checkout order
 * ShopperInsights
     * Add `CustomerSessionRequest.payPalCampaigns` to associate campaigns with a customer session
+
+## 5.31.0 (2026-08-03)
+
+* PayPal
+    * Collect device information to improve PayPal app switch eligibility when it's enabled
+* BraintreeCore
+    * Update Browser Switch version to 3.6.0
 
 ## 5.30.0 (2026-07-21)